+++ feature/home/src/main/java/com/songlib/feature/home/view/HomeScreen.kt
@@ -1,38 +1,21 @@
 package com.songlib.feature.home.view
 
 import android.widget.Toast
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.EditNote
-import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import com.songlib.core.common.entity.UiState
 import com.songlib.core.common.utils.Routes
 import com.songlib.core.data.repos.PrefsRepo
 import com.songlib.core.ui.components.indicators.EmptyState
 import com.songlib.core.ui.components.indicators.ErrorState
-import com.songlib.core.ui.components.listitems.SongSkeletonItem
-import com.songlib.core.ui.components.listitems.shimmerBrush
 import com.songlib.feature.home.HomeViewModel
+import com.songlib.feature.home.components.HomeSkeleton
 import kotlinx.coroutines.flow.collectLatest
 
 @Composable
@@ -59,7 +42,7 @@ fun HomeScreen(
             retryAction = { viewModel.fetchData() }
         )
 
-        UiState.Loading -> HomeSkeletonScreen()
+        UiState.Loading -> HomeSkeleton()
 
         UiState.Filtered -> {
             if (songs.isEmpty()) {
@@ -91,45 +74,3 @@ fun HomeScreen(
     }
 }
 
-@Composable
-private fun HomeSkeletonScreen() {
-    val brush = shimmerBrush()
-    Scaffold { paddingValues ->
-        LazyColumn(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(paddingValues),
-        ) {
-            item {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 10.dp, vertical = 10.dp)
-                        .height(56.dp)
-                        .clip(RoundedCornerShape(12.dp))
-                        .background(brush)
-                )
-            }
-            item {
-                LazyRow(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = 4.dp),
-                    contentPadding = PaddingValues(horizontal = 8.dp),
-                    horizontalArrangement = Arrangement.spacedBy(8.dp),
-                ) {
-                    items(5) {
-                        Box(
-                            modifier = Modifier
-                                .width(72.dp)
-                                .height(28.dp)
-                                .clip(RoundedCornerShape(20.dp))
-                                .background(brush)
-                        )
-                    }
-                }
-            }
-            items(12) { SongSkeletonItem() }
-        }
-    }
-}
