+++ feature/selection/src/main/java/com/songlib/feature/selection/components/SelectionSkeleton.kt
@@ -1,0 +1,66 @@
+package com.songlib.feature.selection.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.songlib.core.ui.components.listitems.SongSkeletonItem
+import com.songlib.core.ui.components.listitems.shimmerBrush
+
+@Preview(showBackground = true)
+@Composable
+fun SelectionSkeleton() {
+    val brush = shimmerBrush()
+    Scaffold { paddingValues ->
+        LazyColumn(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(paddingValues),
+        ) {
+            item {
+                Box(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(horizontal = 10.dp, vertical = 10.dp)
+                        .height(56.dp)
+                        .clip(RoundedCornerShape(12.dp))
+                        .background(brush)
+                )
+            }
+            item {
+                LazyRow(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .padding(vertical = 4.dp),
+                    contentPadding = PaddingValues(horizontal = 8.dp),
+                    horizontalArrangement = Arrangement.spacedBy(8.dp),
+                ) {
+                    items(5) {
+                        Box(
+                            modifier = Modifier
+                                .width(72.dp)
+                                .height(28.dp)
+                                .clip(RoundedCornerShape(20.dp))
+                                .background(brush)
+                        )
+                    }
+                }
+            }
+            items(12) { SongSkeletonItem() }
+        }
+    }
+}
